Refactor: Move admin and transaction classes to separate GUI modules
